+++ include/swift/SILOptimizer/Differentiation/VJPEmitter.h
@@ -18,10 +18,11 @@
 #ifndef SWIFT_SILOPTIMIZER_UTILS_DIFFERENTIATION_VJPEMITTER_H
 #define SWIFT_SILOPTIMIZER_UTILS_DIFFERENTIATION_VJPEMITTER_H
 
-#include "swift/SIL/TypeSubstCloner.h"
+#include "swift/SILOptimizer/Differentiation/DifferentiationInvoker.h"
+#include "swift/SILOptimizer/Differentiation/LinearMapInfo.h"
 #include "swift/SILOptimizer/Analysis/DifferentiableActivityAnalysis.h"
-#include "swift/SILOptimizer/Utils/Differentiation/DifferentiationInvoker.h"
-#include "swift/SILOptimizer/Utils/Differentiation/LinearMapInfo.h"
+
+#include "swift/SIL/TypeSubstCloner.h"
 #include "llvm/ADT/DenseMap.h"
 
 namespace swift {
