+++ lib/SILOptimizer/Differentiation/VJPEmitter.cpp
@@ -17,12 +17,13 @@
 
 #define DEBUG_TYPE "differentiation"
 
-#include "swift/SILOptimizer/Utils/Differentiation/VJPEmitter.h"
+#include "swift/SILOptimizer/Differentiation/VJPEmitter.h"
+#include "swift/SILOptimizer/Differentiation/ADContext.h"
+#include "swift/SILOptimizer/Differentiation/PullbackEmitter.h"
+#include "swift/SILOptimizer/Differentiation/Thunk.h"
+
 #include "swift/SILOptimizer/PassManager/PrettyStackTrace.h"
 #include "swift/SILOptimizer/Utils/CFGOptUtils.h"
-#include "swift/SILOptimizer/Utils/Differentiation/ADContext.h"
-#include "swift/SILOptimizer/Utils/Differentiation/PullbackEmitter.h"
-#include "swift/SILOptimizer/Utils/Differentiation/Thunk.h"
 #include "swift/SILOptimizer/Utils/SILOptFunctionBuilder.h"
 
 namespace swift {
